fix: restructure release notes data
