--- content/changelog/2022-08-30-console-release/index.md
@@ -1,11 +1,0 @@
----
-label: 'Console'
----
-
-### What's new
-
-- UI: Added a feedback form.
-
-### Bug fixes
-
-- UI: Fixed a CORS error for API requests in the Swagger UI. The error occurred when using 'Try it out'.
