+++ content/docs/release-notes.md
@@ -1,0 +1,3 @@
+---
+title: 'Release notes'
+---
