+++ content/release-notes/2022-08-30.md
@@ -1,0 +1,11 @@
+---
+label: 'Console'
+---
+
+### What's new
+
+- UI: Added a feedback form.
+
+### Bug fixes
+
+- UI: Fixed a CORS error for API requests in the Swagger UI. The error occurred when using 'Try it out'.
