+++ gatsby/constants.js
@@ -4,7 +4,7 @@ const DRAFT_FILTER = IS_PRODUCTION ? [false] : [true, false];
 const POST_REQUIRED_FIELDS = ['title', 'description', 'author'];
 const STATIC_PAGE_REQUIRED_FIELDS = ['title'];
 const DOC_REQUIRED_FIELDS = ['title'];
-const CHANGELOG_POST_REQUIRED_FIELDS = ['label'];
+const RELEASE_NOTES_REQUIRED_FIELDS = ['label'];
 
 module.exports = {
   IS_PRODUCTION,
@@ -12,5 +12,5 @@ module.exports = {
   POST_REQUIRED_FIELDS,
   STATIC_PAGE_REQUIRED_FIELDS,
   DOC_REQUIRED_FIELDS,
-  CHANGELOG_POST_REQUIRED_FIELDS,
+  RELEASE_NOTES_REQUIRED_FIELDS,
 };
