+++ gatsby/create-pages.js
@@ -1,7 +1,5 @@
 const createBlogPages = require('./create-blog-pages');
 const createBlogPosts = require('./create-blog-posts');
-const createChangelogPages = require('./create-changelog-pages');
-const createChangelogPosts = require('./create-changelog-posts');
 const createDocPages = require('./create-doc-pages');
 
 // const createStaticPages = require('./create-static-pages');
@@ -11,6 +9,4 @@ module.exports = async (options) => {
   await createBlogPosts(options);
   // await createStaticPages(options);
   await createDocPages(options);
-  await createChangelogPages(options);
-  await createChangelogPosts(options);
 };
