--- src/components/pages/changelog-content/changelog-post-list/index.js
@@ -1,1 +1,0 @@
-export { default } from './changelog-post-list';
