+++ src/components/pages/release-notes/release-note-list/index.js
@@ -1,0 +1,1 @@
+export { default } from './release-note-list';
