--- src/constants/changelog.js
@@ -1,8 +1,0 @@
-const CHANGELOG_BASE_PATH = '/docs/release-notes/';
-const CHANGELOG_POSTS_PER_PAGE = 10;
-
-// We are using ES modules here in order to be able to import variables from this file in gatsby-node.js
-module.exports = {
-  CHANGELOG_BASE_PATH,
-  CHANGELOG_POSTS_PER_PAGE,
-};
